% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digits.R
\name{digits}
\alias{digits}
\title{A function to break a given number n apart into its individual digits}
\usage{
digits(n)
}
\arguments{
\item{n}{A number (preferably an integer).}
}
\value{
All of the digits that compose \code{n}.
}
\description{
A function to break a given number n apart into its individual digits
}
\examples{
digits(10)
}
