% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{A function to return the nth Fibonacci number}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{A number (>= 1).}
}
\value{
The \code{n}th Fibonacci number.
}
\description{
A function to return the nth Fibonacci number
}
\examples{
fibonacci(10)
}
