% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{A function to find the greatest common divistor between two numbers
n and k using Euclid's algorithm}
\usage{
gcd(n, k)
}
\arguments{
\item{n}{The first number (preferably an integer).}

\item{k}{The second number (preferably an integer).}
}
\value{
The greatest common divisor of \code{n} and \code{k}.
}
\description{
A function to find the greatest common divistor between two numbers
n and k using Euclid's algorithm
}
\examples{
gcd(10, 12)
}
