% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{prime_factors}
\alias{prime_factors}
\title{A function to return the prime factorization of a given number as a list}
\usage{
prime_factors(n)
}
\arguments{
\item{n}{A number (preferably an integer).}
}
\value{
The prime factorization of \code{n}.
}
\description{
A function to return the prime factorization of a given number as a list
}
\examples{
prime_factors(10)
}
