% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{primes_below}
\alias{primes_below}
\title{A function to find all prime numbers below a given number n}
\usage{
primes_below(n)
}
\arguments{
\item{n}{A number.}
}
\value{
All of the prime numbers below \code{n}.
}
\description{
A function to find all prime numbers below a given number n
}
\examples{
primes_below(10)
}
